requirements.txt
python-telegram-bot>=20.0
aiosqlite>=0.17.0